- Fixed bug #6203: Disable translations in Keywords does not act as advertised. - Fixed bug: Unused keywords are never deleted. #- Fixed bug: SQL parse error when removing an object containing keywords.
git-svn-id: file:///home/patrick.allaert/svn-git/ezp-repo/ezpublish/stable/3.4@11793 a01eee8c-daba-0310-acae-fa49f3023285

+++ kernel/classes/datatypes/ezkeyword/ezkeywordtype.php
@@ -65,6 +65,22 @@ class eZKeywordType extends eZDataType
     */
     function initializeObjectAttribute( &$contentObjectAttribute, $currentVersion, &$originalContentObjectAttribute )
     {
+        if ( $currentVersion != false )
+        {
+            $originalContentObjectAttributeID = $originalContentObjectAttribute->attribute( 'id' );
+            $contentObjectAttributeID = $contentObjectAttribute->attribute( 'id' );
+
+            // if translating or copying an object
+            if ( $originalContentObjectAttributeID != $contentObjectAttributeID )
+            {
+                // copy keywords links as well
+                $keyword =& $originalContentObjectAttribute->content();
+                if ( is_object( $keyword ) )
+                {
+                    $keyword->store( $contentObjectAttribute );
+                }
+            }
+        }
     }
 
     /*!
@@ -183,7 +199,7 @@ class eZKeywordType extends eZDataType
         $db =& eZDB::instance();
 
         /* First we retrieve all the keyword ID related to this object attribute */
-        $res = $db->arrayQuery( "SELECT keyword_id 
+        $res = $db->arrayQuery( "SELECT keyword_id
                                  FROM ezkeyword_attribute_link
                                  WHERE objectattribute_id='$contentObjectAttributeID'" );
         if ( !count ( $res ) )
@@ -196,7 +212,7 @@ class eZKeywordType extends eZDataType
         foreach ( $res as $record )
             $keywordIDs[] = $record['keyword_id'];
         $keywordIDString = implode( ', ', $keywordIDs );
-        
+
         /* Then we see which ones only have a count of 1 */
         $res = $db->arrayQuery( "SELECT keyword_id
                                  FROM ezkeyword, ezkeyword_attribute_link
@@ -210,7 +226,8 @@ class eZKeywordType extends eZDataType
         $unusedKeywordIDString = implode( ', ', $unusedKeywordIDs );
 
         /* Then we delete those unused keywords */
-        $db->query( "DELETE FROM ezkeyword WHERE id IN ($unusedKeywordIDString)" );
+        if ( $unusedKeywordIDString )
+            $db->query( "DELETE FROM ezkeyword WHERE id IN ($unusedKeywordIDString)" );
 
         /* And as last we remove the link between the keyword and the object
          * attribute to be removed */
